requirements.txt
langgraph
langchain
langchain-openai
fastapi
uvicorn
openai
python-dotenv
pypdf
numpy
pydantic
faiss-cpu